+++ src/test/java/ntnu/idatt2003/group27/models/actions/LadderActionTest.java
@@ -1,17 +1,14 @@
-package ntnu.idatt2003.group27.models;
+package ntnu.idatt2003.group27.models.actions;
 
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
-import javafx.scene.canvas.GraphicsContext;
 import javafx.scene.paint.Color;
-import ntnu.idatt2003.group27.models.actions.LadderAction;
-import ntnu.idatt2003.group27.view.components.LadderCanvas;
+import ntnu.idatt2003.group27.models.Player;
+import ntnu.idatt2003.group27.models.Tile;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
